pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "local-knowledge-agent"
version = "1.0.0"
description = "Streamlitベースのローカル知識エージェント（RAGシステム）"
authors = [{name = "LocalKnowledgeAgent Team"}]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "streamlit>=1.28.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.29",
    "langchain-ollama>=0.1.0",
    "chromadb>=0.4.15",
    "ollama>=0.2.0",
    "PyPDF2>=3.0.1",
    "python-docx>=0.8.11",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
]
